- updated Data Model length fields from integer to double;
- added migration;
- updated db seeds;
- updated webapi Business Model length fields;
- updated `GetMaxCapacityAsync` method parameters;
- added validation functions also for decimal type fields;
- casted compartment dimensions to int when searching for other with same dimensions;
- updated client Business Model length fields;
- reviewed spinedit control style to support decimal values and negative values;
- reviewed all views to change spinedit style;
- recreated webapi contracts.

Also:
- changed view title string;
- minor code cleanings;
- fixed wrong detail view opened with _Show Details_ command;
- fixed compartmentsCount field was never calculated in items controller.

Related work items: #1830

+++ WMS/App/App.Core/Models/Compartment/CompartmentType.cs
@@ -4,9 +4,9 @@
     {
         #region Properties
 
-        public int? Height { get; set; }
+        public double? Height { get; set; }
 
-        public int? Width { get; set; }
+        public double? Width { get; set; }
 
         #endregion
     }
